working on io

+++ src/string.c
@@ -1,8 +1,8 @@
 #include <string.h>
 
-int string_strlen(char* str) {
+size_t string_strlen(char* str) {
     int c = 0;
-    int i = 0;
+    size_t i = 0;
     while (str[c++] != '\0') {
         i++;
     }
